requirements.txt
asgiref
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
django-compressor
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
pycparser==2.20
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
django-cors-headers==3.7.0
django-filter==2.00
django-utils-six==2.0
Pillow
django-hitcount
boto3
django-storages
djangorestframework==3.13.1
django-allauth==0.43.0